requirements.txt
openai==1.54.3
elevenlabs==1.10.0
python-dotenv==1.0.0
pydantic==2.9.2
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.12
sounddevice==0.5.1
soundfile==0.12.1
numpy==1.26.4
pyaudio==0.2.14
pydub==0.25.1
ffmpeg-python==0.2.0
httpx==0.27.2
streamlit==1.54.0
